[Paul] moved the hydra code
